requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appnope==0.1.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
bidict==0.22.1
cachetools==5.3.1
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
comm==0.1.3
contourpy==1.1.0
cycler==0.10.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
eventlet==0.33.3
executing==1.2.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.8.0
hydra-core==1.3.2
idna==2.10
imageio==2.31.1
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1rc0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyBluez @ git+https://github.com/pybluez/pybluez.git@4d46ce14d9e888e3b0c65d5d3ba2a703f8e5e861
pyFltk==1.3.8
Pygments==2.15.1
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyparsing==2.4.7
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
roboflow==1.1.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
supervision==0.10.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
ultralytics==8.0.122
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wget==3.2
wrapt==1.15.0
yarl==1.9.2
